refactor: renaming `hpccache` -> `storagecache`
