+++ internal/services/storagecache/hpc_cache_blob_target_resource.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package hpccache
+package storagecache
 
 import (
 	"fmt"
@@ -15,18 +15,18 @@ import (
 	"github.com/hashicorp/go-azure-sdk/resource-manager/storagecache/2023-05-01/storagetargets"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/tf"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/hpccache/validate"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/services/storagecache/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/validation"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/timeouts"
 )
 
-func resourceHPCCacheBlobNFSTarget() *pluginsdk.Resource {
+func resourceHPCCacheBlobTarget() *pluginsdk.Resource {
 	return &pluginsdk.Resource{
-		Create: resourceHPCCacheBlobNFSTargetCreateUpdate,
-		Read:   resourceHPCCacheBlobNFSTargetRead,
-		Update: resourceHPCCacheBlobNFSTargetCreateUpdate,
-		Delete: resourceHPCCacheBlobNFSTargetDelete,
+		Create: resourceHPCCacheBlobTargetCreateOrUpdate,
+		Update: resourceHPCCacheBlobTargetCreateOrUpdate,
+		Read:   resourceHPCCacheBlobTargetRead,
+		Delete: resourceHPCCacheBlobTargetDelete,
 
 		Importer: pluginsdk.ImporterValidatingResourceId(func(id string) error {
 			_, err := storagetargets.ParseStorageTargetID(id)
@@ -48,14 +48,14 @@ func resourceHPCCacheBlobNFSTarget() *pluginsdk.Resource {
 				ValidateFunc: validate.StorageTargetName,
 			},
 
-			"resource_group_name": commonschema.ResourceGroupName(),
-
 			"cache_name": {
 				Type:         pluginsdk.TypeString,
 				Required:     true,
 				ForceNew:     true,
 				ValidateFunc: validation.StringIsNotEmpty,
 			},
+
+			"resource_group_name": commonschema.ResourceGroupName(),
 
 			"namespace_path": {
 				Type:         pluginsdk.TypeString,
@@ -70,24 +70,6 @@ func resourceHPCCacheBlobNFSTarget() *pluginsdk.Resource {
 				ValidateFunc: commonids.ValidateStorageContainerID,
 			},
 
-			// TODO: use SDK enums once following issue is addressed
-			// https://github.com/Azure/azure-rest-api-specs/issues/13839
-			"usage_model": {
-				Type:     pluginsdk.TypeString,
-				Required: true,
-				ValidateFunc: validation.StringInSlice([]string{
-					"READ_HEAVY_INFREQ",
-					"READ_HEAVY_CHECK_180",
-					"READ_ONLY",
-					"READ_WRITE",
-					"WRITE_WORKLOAD_15",
-					"WRITE_AROUND",
-					"WRITE_WORKLOAD_CHECK_30",
-					"WRITE_WORKLOAD_CHECK_60",
-					"WRITE_WORKLOAD_CLOUDWS",
-				}, false),
-			},
-
 			"access_policy_name": {
 				Type:         pluginsdk.TypeString,
 				Optional:     true,
@@ -98,16 +80,14 @@ func resourceHPCCacheBlobNFSTarget() *pluginsdk.Resource {
 	}
 }
 
-func resourceHPCCacheBlobNFSTargetCreateUpdate(d *pluginsdk.ResourceData, meta interface{}) error {
+func resourceHPCCacheBlobTargetCreateOrUpdate(d *pluginsdk.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).StorageCache.StorageTargets
 	subscriptionId := meta.(*clients.Client).Account.SubscriptionId
 	ctx, cancel := timeouts.ForCreateUpdate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	name := d.Get("name").(string)
-	resourceGroup := d.Get("resource_group_name").(string)
-	cache := d.Get("cache_name").(string)
-	id := storagetargets.NewStorageTargetID(subscriptionId, resourceGroup, cache, name)
+	log.Printf("[INFO] preparing arguments for Azure HPC Cache Blob Target creation.")
+	id := storagetargets.NewStorageTargetID(subscriptionId, d.Get("resource_group_name").(string), d.Get("cache_name").(string), d.Get("name").(string))
 
 	if d.IsNewResource() {
 		resp, err := client.Get(ctx, id)
@@ -118,7 +98,7 @@ func resourceHPCCacheBlobNFSTargetCreateUpdate(d *pluginsdk.ResourceData, meta i
 		}
 
 		if !response.WasNotFound(resp.HttpResponse) {
-			return tf.ImportAsExistsError("azurerm_hpc_cache_blob_nfs_target", id.ID())
+			return tf.ImportAsExistsError("azurerm_hpc_cache_blob_target", id.ID())
 		}
 	}
 
@@ -126,21 +106,18 @@ func resourceHPCCacheBlobNFSTargetCreateUpdate(d *pluginsdk.ResourceData, meta i
 	containerId := d.Get("storage_container_id").(string)
 
 	// Construct parameters
-
 	param := storagetargets.StorageTarget{
 		Properties: &storagetargets.StorageTargetProperties{
 			Junctions: &[]storagetargets.NamespaceJunction{
 				{
 					NamespacePath:   &namespacePath,
 					TargetPath:      pointer.To("/"),
-					NfsExport:       pointer.To("/"),
 					NfsAccessPolicy: pointer.To(d.Get("access_policy_name").(string)),
 				},
 			},
-			TargetType: storagetargets.StorageTargetTypeBlobNfs,
-			BlobNfs: &storagetargets.BlobNfsTarget{
-				Target:     pointer.To(containerId),
-				UsageModel: pointer.To(d.Get("usage_model").(string)),
+			TargetType: storagetargets.StorageTargetTypeClfs,
+			Clfs: &storagetargets.ClfsTarget{
+				Target: pointer.To(containerId),
 			},
 		},
 	}
@@ -151,10 +128,10 @@ func resourceHPCCacheBlobNFSTargetCreateUpdate(d *pluginsdk.ResourceData, meta i
 
 	d.SetId(id.ID())
 
-	return resourceHPCCacheBlobNFSTargetRead(d, meta)
+	return resourceHPCCacheBlobTargetRead(d, meta)
 }
 
-func resourceHPCCacheBlobNFSTargetRead(d *pluginsdk.ResourceData, meta interface{}) error {
+func resourceHPCCacheBlobTargetRead(d *pluginsdk.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).StorageCache.StorageTargets
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
@@ -167,12 +144,12 @@ func resourceHPCCacheBlobNFSTargetRead(d *pluginsdk.ResourceData, meta interface
 	resp, err := client.Get(ctx, *id)
 	if err != nil {
 		if response.WasNotFound(resp.HttpResponse) {
-			log.Printf("[DEBUG] %s was not found - removing from state!", id)
+			log.Printf("[DEBUG] HPC Cache Blob Target %q was not found - removing from state!", id)
 			d.SetId("")
 			return nil
 		}
 
-		return fmt.Errorf("retrieving %s: %+v", id, err)
+		return fmt.Errorf("retrieving HPC Cache Blob Target %q: %+v", id, err)
 	}
 
 	d.Set("name", id.StorageTargetName)
@@ -181,22 +158,19 @@ func resourceHPCCacheBlobNFSTargetRead(d *pluginsdk.ResourceData, meta interface
 
 	if m := resp.Model; m != nil {
 		if props := m.Properties; props != nil {
-			if props.TargetType != storagetargets.StorageTargetTypeBlobNfs {
-				return fmt.Errorf("The type of this HPC Cache Target %s is not a Blob NFS Target", id)
+			if props.TargetType != storagetargets.StorageTargetTypeClfs {
+				return fmt.Errorf("The type of this HPC Cache Target %q is not a Blob Target", id)
 			}
 
 			storageContainerId := ""
-			usageModel := ""
-			if b := props.BlobNfs; b != nil {
-				storageContainerId = pointer.From(b.Target)
-				usageModel = pointer.From(b.UsageModel)
+			if clfs := props.Clfs; clfs != nil && clfs.Target != nil {
+				storageContainerId = *clfs.Target
 			}
 			d.Set("storage_container_id", storageContainerId)
-			d.Set("usage_model", usageModel)
 
 			namespacePath := ""
 			accessPolicy := ""
-			// There is only one namespace path allowed for the blob nfs target,
+			// There is only one namespace path allowed for blob container storage target,
 			// which maps to the root path of it.
 			if props.Junctions != nil && len(*props.Junctions) == 1 && (*props.Junctions)[0].NamespacePath != nil {
 				namespacePath = *(*props.Junctions)[0].NamespacePath
@@ -206,10 +180,11 @@ func resourceHPCCacheBlobNFSTargetRead(d *pluginsdk.ResourceData, meta interface
 			d.Set("access_policy_name", accessPolicy)
 		}
 	}
+
 	return nil
 }
 
-func resourceHPCCacheBlobNFSTargetDelete(d *pluginsdk.ResourceData, meta interface{}) error {
+func resourceHPCCacheBlobTargetDelete(d *pluginsdk.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).StorageCache.StorageTargets
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
@@ -220,7 +195,7 @@ func resourceHPCCacheBlobNFSTargetDelete(d *pluginsdk.ResourceData, meta interfa
 	}
 
 	if err := client.DeleteThenPoll(ctx, *id, storagetargets.DeleteOperationOptions{}); err != nil {
-		return fmt.Errorf("deleting %s: %+v", id, err)
+		return fmt.Errorf("deleting HPC Cache Blob Target %q : %+v", id, err)
 	}
 
 	return nil
