+++ internal/services/storagecache/hpc_cache_blob_target_resource_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package hpccache_test
+package storagecache_test
 
 import (
 	"context"
