+++ internal/services/storagecache/hpc_cache_resource.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package hpccache
+package storagecache
 
 import (
 	"context"
