+++ internal/services/storagecache/managed_lustre_file_system_resource_test.go
@@ -1,4 +1,4 @@
-package hpccache_test
+package storagecache_test
 
 import (
 	"context"
