+++ internal/services/storagecache/validate/cache_namespace_path.go
